// client/src/tools/photos.model.ts
// Generated by https://quicktype.io

export interface PhotosData {
    photos: Photo[];
}

export interface Photo {
    id:       string;
    title:    string;
    caption:  string;
    source:   string;
    comments: Comment[];
}

export interface Comment {
    comment: string;
    author:  string;
}

export interface PhotoProps {
    id:       string;
    title:    string;
    caption:  string;
    source:   string;
    comments: Comment[];
}

export interface JumpProps {
    setCurrentPhoto:Function;
    photos: Photo[];
}

export interface CommentProps {
    refreshJSON:Function;
    setLoading:Function;
    currentPhoto:number;
    photos: Photo[];
}